requirements.txt
django>=4.2.3,<4.3
djangorestframework>=3.14.0,<3.15
drf-spectacular>=0.27,<0.28
PyYAML>=5.1.0,<5.2.0
uritemplate>=3.0.0,<3.1.0
markdown>3.0.0,<3.1.0
Pygments>=2.4.0,<2.5.
Pillow>=10.2.0,<10.3
djangorestframework-simplejwt>=5.3.1,<5.4
psycopg2==2.9.9
aio-pika>=9.4.1,<9.5
